install.sh: Several improvements
HIGHLIGHTING:
* REUSE_FROM

TRAPS:
Bash traps added for ERR and EXIT; they print corresponding
message to stderr.

find_apm_command:
Function returns 0, if it succeedes. If it fails (none
of the `apm`, `apm-beta` and `apm-nightly` commands is
available), returns 1.

LOCAL_APM_COMMAND:
This variable can be set by the parent Bash and it will be respected.
This:

  'LOCAL_APM_COMMAND=apm-custom' ./install

will use the `apm-custom` command instead of one of the `apm`,
`apm-beta` or `apm-nightly`.
Only if the `LOCAL_APM_COMMAND` is unset (or null), the script
falls back to the `find_apm_command` function.